Fix casting issue in psmove_port_sleep_ms
psmove_port_sleep_ms for Windows would fail to sleep because the input
parameter is uint32_t and would cause the operands in ft.QuadPart = -(10
* 1000 * duration_ms) to be promoted to unsigned which caused an
overflow when applying the negative operator resulting in a large
positive value assigned to QuadPart.

+++ src/platform/psmove_port_windows.c
@@ -792,7 +792,7 @@ psmove_port_sleep_ms(uint32_t duration_ms)
     LARGE_INTEGER ft;
 
     // Convert to 100 nanosecond interval, negative value indicates relative time
-    ft.QuadPart = -(10 * 1000 * duration_ms);
+    ft.QuadPart = -(10ll * 1000ll * (LONGLONG)duration_ms);
 
     timer = CreateWaitableTimer(NULL, TRUE, NULL);
     SetWaitableTimer(timer, &ft, 0, NULL, NULL, 0);
